Add support for images with SMask entry

+++ hayro-interpret/src/device.rs
@@ -8,7 +8,7 @@ pub trait Device {
     fn set_paint(&mut self, color: Color);
     fn stroke_path(&mut self, path: &BezPath, stroke_props: &StrokeProps);
     fn fill_path(&mut self, path: &BezPath, fill_props: &FillProps);
-    fn push_layer(&mut self, clip: &BezPath, fill: Fill, opactity: u8);
+    fn push_layer(&mut self, clip: &BezPath, fill: Fill, opactity: f32);
     fn draw_rgba_image(&mut self, image_data: Vec<u8>, width: u32, height: u32);
     fn pop(&mut self);
 }
@@ -31,7 +31,7 @@ pub(crate) enum ReplayInstruction {
     PushLayer {
         clip: BezPath,
         fill: Fill,
-        opacity: u8,
+        opacity: f32,
     },
     PopClip,
 }
